requirements.txt
# Important dependencies
timm
# Already pulled by ComfyUI
numpy
einops
kornia
Pillow
safetensors
# Black listed by manager to avoid side effects
# torch
# torchvision
